backend/requirements.txt
absl-py==2.4.0
astunparse==1.6.3
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
comtypes==1.4.16
deepface==0.0.99
dnspython==2.8.0
et_xmlfile==2.0.0
filelock==3.25.2
fire==0.7.1
Flask==3.1.3
Flask-Bcrypt==1.0.1
flask-cors==6.0.2
flatbuffers==25.12.19
gast==0.7.0
gdown==5.2.1
google-pasta==0.2.0
grpcio==1.80.0
gunicorn==25.3.0
h5py==3.14.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
keras==3.14.0
libclang==18.1.1
lightdsa==0.0.3
lightecc==0.0.5
lightphe==0.0.22
lz4==4.4.5
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
mtcnn==1.0.0
namex==0.1.0
numpy==2.4.4
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pandas==3.0.2
pillow==12.2.0
protobuf==7.34.1
Pygments==2.20.0
pymongo==4.16.0
pypiwin32==223
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pyttsx3==2.99
pywin32==311
requests==2.33.1
retina-face==0.0.17
rich==14.3.3
scipy==1.16.2
setuptools==82.0.1
six==1.17.0
soupsieve==2.8.3
sympy==1.14.0
tensorflow==2.21.0
tf-keras==2.21.0
termcolor==3.3.0
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
Werkzeug==3.1.8
wheel==0.46.3
wrapt==2.1.2
